requirements.txt
discord.py[voice]
python-dotenv
ollama
huggingface-hub
setuptools
opencv-python
numpy
pandas
pillow
tqdm
jsonlines
google-genai